fix room join for freenode channels
I had this problem that hubot wouldn't join rooms and just idled connected to
the freenode servers. I tracked it down to the `NOTICE` message parsing. The
freenode servers would not send 'Password accepted' but rather something like
'You are now identified for ...'. I added a condition to take that into
account.