Add support for emotes
I'm not sure this solution fits in the hubot-irc core, as it disrupts
how people would currently use hubot-irc to track regular messages and
ACTION messages together. If people relied receiving both as
TextMessages, people will stop receiving emotes in their listners. That
said, it feels right as a solution. That, or simply adding a property
that denotes that the TextMessage object is an emote.